moved filterPanel text to /projects/displayTexts
